kernel: core_hook: harden prctl handler
detection is done by comparing a 0xDEADBEEF call to a non-0xDEADBEEF one.
which yeah, you will see that the non-0xDEADBEEF one returns early.

yes I know this causes delays for all prctl calls, as we straight up check uid,
but this keeps the delay consistent, which is what we want.

another is that we only perform this recrowning if no manager is crowned
theres really no need for this if manager already exists

- temp fix for cimb octo's prctl abuse

Signed-off-by: backslashxx <118538522+backslashxx@users.noreply.github.com>